+++ src/js/state/index.mjs
@@ -31,6 +31,11 @@ import {
 } from '../lib/utils-html.mjs'
 
 import {
+  HOOK_SYNCNOW,
+  triggerEvent
+} from '../lib/events.mjs'
+
+import {
   DEMO_MODE,
   UnexpectedStatus,
   apiGetTable,
@@ -52,10 +57,6 @@ import {
 } from '../api/index.mjs'
 
 import {
-  syncNow
-} from '../view/room/sync.mjs'
-
-import {
   runError,
   apiError
 } from '../view/error/index.mjs'
@@ -68,6 +69,7 @@ import {
   findPiece,
   populatePiecesDefaults,
   populateSetupDefaults,
+  populateLibraryDefaults,
   clampToTableSize,
   nameToLayer,
   getRoomMediaURL,
@@ -146,14 +148,14 @@ export function setTableNo (no, sync = true) {
   if (no >= 1 && no <= 9) {
     tableNo = no
     setRoomPreference(PREFS.TABLE, tableNo)
-    if (sync) syncNow(true)
+    if (sync) triggerEvent(HOOK_SYNCNOW, true)
   }
 }
 
 /**
  * Get (cached) state for a given slot/table.
  *
- * @param {number} no Table slot 0..9. Defaults to current one.
+ * @param {number} no Table slot 1..9. Defaults to current one.
  * @returns {object} Table array.
  */
 export function getTable (no = getTableNo()) {
@@ -436,7 +438,7 @@ export function loadRoom (name, t) {
 export function patchSetup (setup, sync = true) {
   return apiPatchSetup(room.name, setup, getToken())
     .catch(error => apiError(error, room.name, [404]))
-    .finally(() => { if (sync) syncNow() })
+    .finally(() => { if (sync) triggerEvent(HOOK_SYNCNOW) })
 }
 
 /**
@@ -459,7 +461,7 @@ export function addRoom (room, snapshot) {
  */
 export function undo (no = getTableNo(), sync = true) {
   return apiPostUndo(room.name, no, getToken())
-    .finally(() => { if (sync) syncNow() })
+    .finally(() => { if (sync) triggerEvent(HOOK_SYNCNOW) })
 }
 
 /**
@@ -645,7 +647,7 @@ export function deletePiece (pieceId, sync = true) {
   if (findPiece(pieceId)?.f & FLAGS.NO_DELETE) return Promise.resolve() // can't delete those
   return apiDeletePiece(room.name, getTableNo(), pieceId, getToken())
     .catch(error => apiError(error, room.name))
-    .finally(() => { if (sync) syncNow() })
+    .finally(() => { if (sync) triggerEvent(HOOK_SYNCNOW) })
 }
 
 /**
@@ -664,7 +666,7 @@ export function deletePieces (pieceIds, sync = true) {
   }
   return apiDeletePieces(room.name, getTableNo(), pieceIds, getToken())
     .catch(error => apiError(error, room.name))
-    .finally(() => { if (sync) syncNow() })
+    .finally(() => { if (sync) triggerEvent(HOOK_SYNCNOW) })
 }
 
 /**
@@ -680,7 +682,7 @@ export function deletePieces (pieceIds, sync = true) {
 export function updateAsset (asset, sync = true) {
   return apiPatchAsset(room.name, asset, getToken())
     .catch(error => apiError(error, room.name))
-    .finally(() => { if (sync) syncNow() })
+    .finally(() => { if (sync) triggerEvent(HOOK_SYNCNOW) })
 }
 
 /**
@@ -697,7 +699,7 @@ export function updateAsset (asset, sync = true) {
 export function deleteAsset (assetId, sync = true) {
   return apiDeleteAsset(room.name, assetId, getToken())
     .catch(error => apiError(error, room.name))
-    .finally(() => { if (sync) syncNow() })
+    .finally(() => { if (sync) triggerEvent(HOOK_SYNCNOW) })
 }
 
 /**
@@ -712,7 +714,7 @@ export function deleteAsset (assetId, sync = true) {
 export function updateTable (table, sync = true) {
   return apiPutTable(room.name, getTableNo(), table, getToken())
     .catch(error => apiError(error, room.name))
-    .finally(() => { if (sync) syncNow() })
+    .finally(() => { if (sync) triggerEvent(HOOK_SYNCNOW) })
 }
 
 /**
@@ -759,7 +761,7 @@ export function createPieces (pieces, select = false, sync = true) {
       return ids
     })
     .catch(error => apiError(error, room.name))
-    .finally(() => { if (sync) syncNow() })
+    .finally(() => { if (sync) triggerEvent(HOOK_SYNCNOW) })
 }
 
 /**
@@ -828,7 +830,7 @@ export function isTabActive () {
  */
 export function setTabActive (state) {
   tabActive = state
-  if (state && room) syncNow()
+  if (state && room) triggerEvent(HOOK_SYNCNOW)
 }
 
 // --- internal ----------------------------------------------------------------
@@ -862,6 +864,7 @@ export function setTable (no, data) {
 export function setRoom (data) {
   if (data) {
     data.setup = populateSetupDefaults(data.setup)
+    data.library = populateLibraryDefaults(data.library)
   }
   room = data
 }
@@ -898,7 +901,7 @@ function patchPiece (pieceId, patch, sync = true) {
   if (patch.l) patch.l = nameToLayer(patch.l)
   return apiPatchPiece(room.name, getTableNo(), pieceId, patch, getToken())
     .catch(error => apiError(error, room.name, [404]))
-    .finally(() => { if (sync) syncNow() })
+    .finally(() => { if (sync) triggerEvent(HOOK_SYNCNOW) })
 }
 
 /**
@@ -917,27 +920,8 @@ function patchPieces (patches, sync = true) {
   }
   return apiPatchPieces(room.name, getTableNo(), sane, getToken())
     .catch(error => apiError(error, room.name, [404]))
-    .finally(() => { if (sync) syncNow() })
+    .finally(() => { if (sync) triggerEvent(HOOK_SYNCNOW) })
 }
-
-// /**
-//  * Create a piece on the server.
-//  *
-//  * @param {object} piece The full piece to send to the server.
-//  * @param {boolean} select Optional. If false (default), piece will not get selected.
-//  * @param {boolean} sync Optional. If true (default), trigger table sync.
-//  * @returns {object} Promise of the ID of the new piece.
-//  */
-// function createPiece (piece, select = false, sync = true) {
-//   if (piece.l) piece.l = nameToLayer(piece.l)
-//   return apiPostPiece(room.name, getTableNo(), stripPiece(piece), getToken())
-//     .then(piece => {
-//       if (select) selectionAdd(piece.id, true)
-//       return piece.id
-//     })
-//     .catch(error => apiError(error, room.name))
-//     .finally(() => { if (sync) syncNow() })
-// }
 
 export const _test = {
   setTable,