+++ test/integration/api/api-assets.mjs
@@ -66,6 +66,7 @@ function testApiCreateAsset (api, version, room) {
       format: 'jpg',
       h: 2,
       w: 3,
+      d: 4,
       type: LAYER_TILE,
       tx: 'wood',
       name: 'upload.test'
@@ -73,9 +74,11 @@ function testApiCreateAsset (api, version, room) {
   }, body => {
     expect(body).to.be.an('object')
     expect(body.bg).to.be.eql('#808080')
-    expect(body.format).to.be.eql('jpg')
+    expect(body.format).to.be.eql(undefined)
+    expect(body.base64).to.be.eql(undefined)
     expect(body.h).to.be.eql(2)
     expect(body.w).to.be.eql(3)
+    expect(body.d).to.be.eql(4)
     expect(body.type).to.be.eql(LAYER_TILE)
     expect(body.name).to.be.eql('upload.test')
     expect(body.tx).to.be.eql('wood')
@@ -91,17 +94,62 @@ function testApiCreateAsset (api, version, room) {
 
     expect(body.library.tile[index].id).to.be.an('string')
     expect(body.library.tile[index].media).to.be.an('array')
-    expect(body.library.tile[index].media[0]).to.be.eql('upload.test.3x2x1.808080.wood.jpg')
+    expect(body.library.tile[index].media[0]).to.be.eql('upload.test.3x2x1x4.808080.wood.jpg')
     expect(body.library.tile[index].bg).to.be.eql('#808080')
     expect(body.library.tile[index].h).to.be.eql(2)
     expect(body.library.tile[index].w).to.be.eql(3)
+    expect(body.library.tile[index].d).to.be.eql(4)
     expect(body.library.tile[index].type).to.be.eql(LAYER_TILE)
     expect(body.library.tile[index].name).to.be.eql('upload.test')
     expect(body.library.tile[index].tx).to.be.eql('wood')
   }, 200)
 
+  // upload another asset
+  testJsonPost(api, () => `/rooms/${room}/assets/`, () => {
+    return {
+      base64: Buffer.from(image).toString('base64'),
+      bg: '#808080',
+      format: 'jpg',
+      h: 2,
+      w: 3,
+      d: 2, // default depth for tiles
+      type: LAYER_TILE,
+      tx: 'wood',
+      name: 'upload.test2'
+    }
+  }, body => {
+    expect(body).to.be.an('object')
+    expect(body.bg).to.be.eql('#808080')
+    expect(body.h).to.be.eql(2)
+    expect(body.w).to.be.eql(3)
+    expect(body.d).to.be.eql(undefined)
+    expect(body.type).to.be.eql(LAYER_TILE)
+    expect(body.name).to.be.eql('upload.test2')
+    expect(body.tx).to.be.eql('wood')
+  }, 201)
+
+  // library must contain asset now
+  testJsonGet(api, () => `/rooms/${room}/`, body => {
+    expect(body).to.be.an('object')
+    expect(body.library).to.be.an('object')
+    expect(body.library.tile.length).to.be.eql(data.tile.length + 2)
+
+    const index = body.library.tile.length - 1
+
+    expect(body.library.tile[index].id).to.be.an('string')
+    expect(body.library.tile[index].media).to.be.an('array')
+    expect(body.library.tile[index].media[0]).to.be.eql('upload.test2.3x2x1.808080.wood.jpg')
+    expect(body.library.tile[index].bg).to.be.eql('#808080')
+    expect(body.library.tile[index].h).to.be.eql(2)
+    expect(body.library.tile[index].w).to.be.eql(3)
+    expect(body.library.tile[index].d).to.be.eql(undefined)
+    expect(body.library.tile[index].type).to.be.eql(LAYER_TILE)
+    expect(body.library.tile[index].name).to.be.eql('upload.test2')
+    expect(body.library.tile[index].tx).to.be.eql('wood')
+  }, 200)
+
   // check asset blob
-  testGetBuffer(api, () => `/data/rooms/${room}/assets/tile/upload.test.3x2x1.808080.wood.jpg`, () => {}, (buffer) => {
+  testGetBuffer(api, () => `/data/rooms/${room}/assets/tile/upload.test2.3x2x1.808080.wood.jpg`, () => {}, (buffer) => {
     expect(buffer.toString('utf-8')).to.be.eql(image)
   }, 200)
 
@@ -116,7 +164,7 @@ function testApiUpdateAssetOtherBaseMask (api, version, room) {
     expect(body.library.other.length).to.be.eql(_.other + classic.other)
   }, 200)
 
-  // patch name < 10 sides
+  // patch name < 10 sides - _.d4Light
   testJsonPatch(api, () => `/rooms/${room}/assets/j8hqf000/`, () => {
     return {
       id: 'j8hqf000',
@@ -127,18 +175,18 @@ function testApiUpdateAssetOtherBaseMask (api, version, room) {
     expect(body.name).to.be.eql('barFoo.fooBar')
     expect(body.id).to.be.eql('EJwm7300')
     expect(body.type).to.be.eql('other')
-    expect(body.w).to.be.eql(1)
-    expect(body.h).to.be.eql(1)
+    expect(body.w).to.be.eql(undefined)
+    expect(body.h).to.be.eql(undefined)
     expect(body.bg).to.be.eql('2')
     expect(body.tx).to.be.eql('paper')
-    expect(body.base).to.be.eql('barFoo.fooBar.1x1x0.2.paper.png')
-    expect(body.mask).to.be.eql('barFoo.fooBar.1x1xX.2.paper.svg')
+    expect(body.base).to.be.eql('barFoo.fooBar.1x1x0x5.2.paper.png')
+    expect(body.mask).to.be.eql('barFoo.fooBar.1x1xXx5.2.paper.svg')
     for (const media of body.media) {
-      expect(media.match(/barFoo.fooBar\.1x1x[0-9]\.2\.paper\.svg/)).to.be.an('array')
+      expect(media.match(/barFoo.fooBar\.1x1x[0-9]x5\.2\.paper\.svg/)).to.be.an('array')
     }
   }, 200)
 
-  // patch name >= 10 sides
+  // patch name >= 10 sides - _.d12Dark
   testJsonPatch(api, () => `/rooms/${room}/assets/v8Vvg200/`, () => {
     return {
       id: 'v8Vvg200',
@@ -149,25 +197,25 @@ function testApiUpdateAssetOtherBaseMask (api, version, room) {
     expect(body.name).to.be.eql('fooBar.fooBar')
     expect(body.id).to.be.eql('uwlgt100')
     expect(body.type).to.be.eql('other')
-    expect(body.w).to.be.eql(1)
-    expect(body.h).to.be.eql(1)
+    expect(body.w).to.be.eql(undefined)
+    expect(body.h).to.be.eql(undefined)
     expect(body.bg).to.be.eql('1')
     expect(body.tx).to.be.eql('paper')
-    expect(body.base).to.be.eql('fooBar.fooBar.1x1x00.1.paper.png')
-    expect(body.mask).to.be.eql('fooBar.fooBar.1x1xXX.1.paper.svg')
+    expect(body.base).to.be.eql('fooBar.fooBar.1x1x00x5.1.paper.png')
+    expect(body.mask).to.be.eql('fooBar.fooBar.1x1xXXx5.1.paper.svg')
     expect(body.media).to.be.eql([
-      'fooBar.fooBar.1x1x01.1.paper.svg',
-      'fooBar.fooBar.1x1x02.1.paper.svg',
-      'fooBar.fooBar.1x1x03.1.paper.svg',
-      'fooBar.fooBar.1x1x04.1.paper.svg',
-      'fooBar.fooBar.1x1x05.1.paper.svg',
-      'fooBar.fooBar.1x1x06.1.paper.svg',
-      'fooBar.fooBar.1x1x07.1.paper.svg',
-      'fooBar.fooBar.1x1x08.1.paper.svg',
-      'fooBar.fooBar.1x1x09.1.paper.svg',
-      'fooBar.fooBar.1x1x10.1.paper.svg',
-      'fooBar.fooBar.1x1x11.1.paper.svg',
-      'fooBar.fooBar.1x1x12.1.paper.svg'
+      'fooBar.fooBar.1x1x01x5.1.paper.svg',
+      'fooBar.fooBar.1x1x02x5.1.paper.svg',
+      'fooBar.fooBar.1x1x03x5.1.paper.svg',
+      'fooBar.fooBar.1x1x04x5.1.paper.svg',
+      'fooBar.fooBar.1x1x05x5.1.paper.svg',
+      'fooBar.fooBar.1x1x06x5.1.paper.svg',
+      'fooBar.fooBar.1x1x07x5.1.paper.svg',
+      'fooBar.fooBar.1x1x08x5.1.paper.svg',
+      'fooBar.fooBar.1x1x09x5.1.paper.svg',
+      'fooBar.fooBar.1x1x10x5.1.paper.svg',
+      'fooBar.fooBar.1x1x11x5.1.paper.svg',
+      'fooBar.fooBar.1x1x12x5.1.paper.svg'
     ])
   }, 200)
 
@@ -286,9 +334,35 @@ function testApiUpdateAssetOverlayMaterial (api, version, room) {
     expect(body.id).to.be.eql('Jvqgu300')
     expect(body.type).to.be.eql('overlay')
     expect(body.w).to.be.eql(3)
-    expect(body.h).to.be.eql(3)
+    expect(body.h).to.be.eql(undefined)
     expect(body.bg).to.be.eql('2')
     expect(body.tx).to.be.eql(undefined)
+    expect(body.base).to.be.eql(undefined)
+    expect(body.mask).to.be.eql('_.fooBar.3x3xX.2.svg')
+    expect(body.media).to.be.eql([])
+  }, 200)
+
+  // patch d - non-default
+  testJsonPatch(api, () => `/rooms/${room}/assets/Jvqgu300/`, () => {
+    return {
+      id: 'kVTKu200',
+      d: 2
+    }
+  }, body => {
+    expect(body.d).to.be.eql(2)
+    expect(body.base).to.be.eql(undefined)
+    expect(body.mask).to.be.eql('_.fooBar.3x3xXx2.2.svg')
+    expect(body.media).to.be.eql([])
+  }, 200)
+
+  // patch d - default
+  testJsonPatch(api, () => `/rooms/${room}/assets/Jvqgu300/`, () => {
+    return {
+      id: 'kVTKu200',
+      d: 0
+    }
+  }, body => {
+    expect(body.d).to.be.eql(undefined)
     expect(body.base).to.be.eql(undefined)
     expect(body.mask).to.be.eql('_.fooBar.3x3xX.2.svg')
     expect(body.media).to.be.eql([])
@@ -569,8 +643,9 @@ function testApiUpdateAssetToken (api, version, room) {
     expect(body.name).to.be.eql('blah')
     expect(body.id).to.be.eql('rh0-6200')
     expect(body.type).to.be.eql('token')
-    expect(body.w).to.be.eql(1)
-    expect(body.h).to.be.eql(1)
+    expect(body.w).to.be.eql(undefined) // default
+    expect(body.h).to.be.eql(undefined) // default
+    expect(body.d).to.be.eql(undefined) // default
     expect(body.bg).to.be.eql('1')
     expect(body.tx).to.be.eql('wood')
     expect(body.base).to.be.eql(undefined)
@@ -592,6 +667,7 @@ function testApiUpdateAssetToken (api, version, room) {
     expect(body.type).to.be.eql('token')
     expect(body.w).to.be.eql(5)
     expect(body.h).to.be.eql(1)
+    expect(body.d).to.be.eql(undefined) // default
     expect(body.bg).to.be.eql('1')
     expect(body.tx).to.be.eql('wood')
     expect(body.base).to.be.eql(undefined)
@@ -613,12 +689,57 @@ function testApiUpdateAssetToken (api, version, room) {
     expect(body.type).to.be.eql('token')
     expect(body.w).to.be.eql(5)
     expect(body.h).to.be.eql(3)
+    expect(body.d).to.be.eql(undefined) // default
     expect(body.bg).to.be.eql('1')
     expect(body.tx).to.be.eql('wood')
     expect(body.base).to.be.eql(undefined)
     expect(body.mask).to.be.eql(undefined)
     for (const media of body.media) {
       expect(media.match(/blah\.5x3x[0-9][0-9]\.1\.wood\.svg/)).to.be.an('array')
+    }
+  }, 200)
+
+  // patch d - default
+  testJsonPatch(api, () => `/rooms/${room}/assets/HRXAK100/`, () => {
+    return {
+      id: 'HRXAK100',
+      d: 2
+    }
+  }, body => {
+    expect(body.name).to.be.eql('blah')
+    expect(body.id).to.be.eql('HRXAK100')
+    expect(body.type).to.be.eql('token')
+    expect(body.w).to.be.eql(5)
+    expect(body.h).to.be.eql(3)
+    expect(body.d).to.be.eql(undefined)
+    expect(body.bg).to.be.eql('1')
+    expect(body.tx).to.be.eql('wood')
+    expect(body.base).to.be.eql(undefined)
+    expect(body.mask).to.be.eql(undefined)
+    for (const media of body.media) {
+      expect(media.match(/blah\.5x3x[0-9][0-9]\.1\.wood\.svg/)).to.be.an('array')
+    }
+  }, 200)
+
+  // patch d - non-default
+  testJsonPatch(api, () => `/rooms/${room}/assets/HRXAK100/`, () => {
+    return {
+      id: 'HRXAK100',
+      d: 8
+    }
+  }, body => {
+    expect(body.name).to.be.eql('blah')
+    expect(body.id).to.be.eql('HRXAK100')
+    expect(body.type).to.be.eql('token')
+    expect(body.w).to.be.eql(5)
+    expect(body.h).to.be.eql(3)
+    expect(body.d).to.be.eql(8)
+    expect(body.bg).to.be.eql('1')
+    expect(body.tx).to.be.eql('wood')
+    expect(body.base).to.be.eql(undefined)
+    expect(body.mask).to.be.eql(undefined)
+    for (const media of body.media) {
+      expect(media.match(/blah\.5x3x[0-9][0-9]x8\.1\.wood\.svg/)).to.be.an('array')
     }
   }, 200)
 
@@ -634,12 +755,13 @@ function testApiUpdateAssetToken (api, version, room) {
     expect(body.type).to.be.eql('token')
     expect(body.w).to.be.eql(5)
     expect(body.h).to.be.eql(3)
+    expect(body.d).to.be.eql(8)
     expect(body.bg).to.be.eql('1')
     expect(body.tx).to.be.eql('linen')
     expect(body.base).to.be.eql(undefined)
     expect(body.mask).to.be.eql(undefined)
     for (const media of body.media) {
-      expect(media.match(/blah\.5x3x[0-9][0-9]\.1\.linen\.svg/)).to.be.an('array')
+      expect(media.match(/blah\.5x3x[0-9][0-9]x8\.1\.linen\.svg/)).to.be.an('array')
     }
   }, 200)
 
@@ -655,12 +777,13 @@ function testApiUpdateAssetToken (api, version, room) {
     expect(body.type).to.be.eql('token')
     expect(body.w).to.be.eql(5)
     expect(body.h).to.be.eql(3)
+    expect(body.d).to.be.eql(8)
     expect(body.bg).to.be.eql('3')
     expect(body.tx).to.be.eql('linen')
     expect(body.base).to.be.eql(undefined)
     expect(body.mask).to.be.eql(undefined)
     for (const media of body.media) {
-      expect(media.match(/blah\.5x3x[0-9][0-9]\.3\.linen\.svg/)).to.be.an('array')
+      expect(media.match(/blah\.5x3x[0-9][0-9]x8\.3\.linen\.svg/)).to.be.an('array')
     }
   }, 200)
 
@@ -675,12 +798,13 @@ function testApiUpdateAssetToken (api, version, room) {
     expect(newAsset.type).to.be.eql('token')
     expect(newAsset.w).to.be.eql(5)
     expect(newAsset.h).to.be.eql(3)
+    expect(newAsset.d).to.be.eql(8)
     expect(newAsset.bg).to.be.eql('3')
     expect(newAsset.tx).to.be.eql('linen')
     expect(newAsset.base).to.be.eql(undefined)
     expect(newAsset.mask).to.be.eql(undefined)
     for (const media of newAsset.media) {
-      expect(media.match(/blah\.5x3x[0-9][0-9]\.3\.linen\.svg/)).to.be.an('array')
+      expect(media.match(/blah\.5x3x[0-9][0-9]x8\.3\.linen\.svg/)).to.be.an('array')
     }
   }, 200)